+++ src/evaluation.rs
@@ -1,22 +1,20 @@
 use chess::*;
-use features::Model;
 use mcts::{Evaluator, GameState};
 use policy_features::evaluate_moves;
 use search::{GooseMCTS, SCALE};
 use shakmaty::Position;
 use shakmaty_syzygy::Wdl;
+use state;
 use state::{MoveList, Player, State};
 use tablebase::probe_tablebase_wdl;
 
 const MATE_FACTOR: f32 = 1.1;
 
-pub struct GooseEval {
-    model: Model,
-}
+pub struct GooseEval;
 
 impl GooseEval {
-    pub fn new(model: Model) -> Self {
-        Self { model }
+    pub fn new() -> Self {
+        Self
     }
 }
 
@@ -46,7 +44,7 @@ impl Evaluator<GooseMCTS> for GooseEval {
                 _ => 0,
             }
         } else {
-            (self.model.score(state, &features) * SCALE as f32) as i64
+            (evaluate_state(state, &features) * SCALE as f32) as i64
         };
         (move_evaluations, state_evaluation, tb_hit)
     }
@@ -57,4 +55,51 @@ impl Evaluator<GooseMCTS> for GooseEval {
             Color::Black => -*evaln,
         }
     }
+}
+
+const NUMBER_INPUTS: usize = state::NUMBER_FEATURES;
+const NUMBER_HIDDEN: usize = 128;
+const NUMBER_OUTPUTS: usize = 1;
+
+#[allow(clippy::excessive_precision)]
+const EVAL_HIDDEN_BIAS: [f32; NUMBER_HIDDEN] = include!("model/hidden_bias_0");
+
+#[allow(clippy::excessive_precision)]
+const EVAL_HIDDEN_WEIGHTS: [[f32; NUMBER_INPUTS]; NUMBER_HIDDEN] =
+    include!("model/hidden_weights_0");
+
+#[allow(clippy::excessive_precision)]
+const EVAL_OUTPUT_WEIGHTS: [[f32; NUMBER_HIDDEN]; NUMBER_OUTPUTS] =
+    include!("model/output_weights");
+
+fn evaluate_state(state: &State, features: &[f32; state::NUMBER_FEATURES]) -> f32 {
+    #[allow(clippy::uninit_assumed_init)]
+    let mut hidden_layer: [f32; NUMBER_HIDDEN] =
+        unsafe { std::mem::MaybeUninit::uninit().assume_init() };
+
+    hidden_layer.copy_from_slice(&EVAL_HIDDEN_BIAS);
+
+    for i in 0..features.len() {
+        if features[i] > 0.5 {
+            for j in 0..hidden_layer.len() {
+                hidden_layer[j] += EVAL_HIDDEN_WEIGHTS[j][i];
+            }
+        }
+    }
+
+    let mut result = 0.;
+
+    let weights = EVAL_OUTPUT_WEIGHTS[0];
+
+    for i in 0..hidden_layer.len() {
+        result += weights[i] * hidden_layer[i].max(0.);
+    }
+
+    result = result.tanh();
+
+    if state.board().side_to_move() == Color::Black {
+        result = -result;
+    }
+
+    result
 }
