+++ src/uci.rs
@@ -60,7 +60,6 @@ pub fn main(commands: Vec<String>) {
                 "stop"       => search = search.stop_and_print(),
                 "quit"       => return,
                 "go"         => search = search.go(tokens, &sender),
-                "eval"       => search.print_eval(),
                 "movelist"   => search.print_move_list(),
                 "sizelist"   => print_size_list(),
                 "info"       => search.print_info(),
